requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
aiofiles
passlib
passlib[bcrypt]
python-jose[cryptography]
bcrypt